# cybercrime/openapi.yaml
components:
  schemas:
...
    FreezeWalletRequest:
      properties:
        freeze_reason:
          title: Freeze Reason
          type: string
        frozen_by:
          title: Frozen By
          type: string
      required:
      - freeze_reason
      - frozen_by
      title: FreezeWalletRequest
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    UnfreezeWalletRequest:
      properties:
        unfreeze_reason:
          title: Unfreeze Reason
          type: string
        unfrozen_by:
          title: Unfrozen By
          type: string
      required:
      - unfreeze_reason
      - unfrozen_by
      title: UnfreezeWalletRequest
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    WalletCreate:
      properties:
        address:
          title: Address
          type: string
        anchor_status:
          anyOf:
          - type: string
          - type: 'null'
          default: pending
          title: Anchor Status
        balance:
          anyOf:
          - type: string
          - type: 'null'
          title: Balance
        case_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Case Id
        label:
          anyOf:
          - type: string
          - type: 'null'
          title: Label
        risk_level:
          anyOf:
          - type: string
          - type: 'null'
          default: low
          title: Risk Level
      required:
      - address
      title: WalletCreate
      type: object
    WalletResponse:
      properties:
        address:
          title: Address
          type: string
        anchor_status:
          anyOf:
          - type: string
          - type: 'null'
          default: pending
          title: Anchor Status
        balance:
          anyOf:
          - type: string
          - type: 'null'
          title: Balance
        case_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Case Id
        created_at:
          format: date-time
          title: Created At
          type: string
        id:
          title: Id
          type: integer
        label:
          anyOf:
          - type: string
          - type: 'null'
          title: Label
        last_activity:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Last Activity
        risk_level:
          anyOf:
          - type: string
          - type: 'null'
          default: low
          title: Risk Level
      required:
      - address
      - label
      - balance
      - risk_level
      - anchor_status
      - id
      - case_id
      - last_activity
      - created_at
      title: WalletResponse
      type: object
...
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
      type: oauth2
info:
  description: Backend API for cybercrime investigation dashboard
  title: Cybercrime Investigation API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/v1/wallets/:
    get:
      description: Get all wallets under investigation
      operationId: get_wallets_api_v1_wallets__get
      parameters:
      - in: query
        name: case_id
        required: false
        schema:
          title: Case Id
          type: integer
      - in: query
        name: skip
        required: false
        schema:
          default: 0
          title: Skip
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/WalletResponse'
                title: Response Get Wallets Api V1 Wallets  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Wallets
      tags:
      - wallets
    post:
      description: Create a new wallet entry
      operationId: create_wallet_api_v1_wallets__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Wallet
      tags:
      - wallets
  /api/v1/wallets/frozen/list:
    get:
      description: Get all frozen wallets
      operationId: get_frozen_wallets_api_v1_wallets_frozen_list_get
      parameters:
      - in: query
        name: skip
        required: false
        schema:
          default: 0
          title: Skip
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Frozen Wallets
      tags:
      - wallets
...
  /api/v1/wallets/search/{wallet_address}:
    get:
      description: Search wallet by address and get all related data
      operationId: search_wallet_api_v1_wallets_search__wallet_address__get
      parameters:
      - in: path
        name: wallet_address
        required: true
        schema:
          title: Wallet Address
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Search Wallet
      tags:
      - wallets
  /api/v1/wallets/unfrozen/list:
    get:
      description: Get all unfrozen wallets (previously frozen)
      operationId: get_unfrozen_wallets_api_v1_wallets_unfrozen_list_get
      parameters:
      - in: query
        name: skip
        required: false
        schema:
          default: 0
          title: Skip
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Unfrozen Wallets
      tags:
      - wallets
...
  /api/v1/wallets/{wallet_id}:
    get:
      description: Get a specific wallet by ID
      operationId: get_wallet_api_v1_wallets__wallet_id__get
      parameters:
      - in: path
        name: wallet_id
        required: true
        schema:
          title: Wallet Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Wallet
      tags:
      - wallets
  /api/v1/wallets/{wallet_id}/freeze:
    post:
      description: Freeze a wallet
      operationId: freeze_wallet_api_v1_wallets__wallet_id__freeze_post
      parameters:
      - in: path
        name: wallet_id
        required: true
        schema:
          title: Wallet Id
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FreezeWalletRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Freeze Wallet
      tags:
      - wallets
  /api/v1/wallets/{wallet_id}/unfreeze:
    post:
      description: Unfreeze a wallet
      operationId: unfreeze_wallet_api_v1_wallets__wallet_id__unfreeze_post
      parameters:
      - in: path
        name: wallet_id
        required: true
        schema:
          title: Wallet Id
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnfreezeWalletRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Unfreeze Wallet
      tags:
      - wallets
...